Add sge_pipeline package for top-level SGE pipeline
Currently this is just the ScratchDistributor module for copying
blast_db_dir to all the compute nodes.

```
luigi --module eppicpipeline.luigi.sge_pipeline ScratchDistributor --seed=`date +%s`
```

Note the use of `--seed` to force rsync to run even if there are existing
files. This always needs to be passed at the command line. It also adds
a global param `sentinel_dir` for flagging completed tasks.

The package ships with a merlin4-specific list of compute nodes, but
this can be overwritten with the `--host-file` param.